cleaning up and making some tests
