--- frontend/libs/studio-components/src/components/StudioCheckboxTable/StudioCheckboxTable.module.css
@@ -1,3 +1,0 @@
-.header {
-  background-color: var(--fds-colors-grey-100);
-}
