+++ frontend/libs/studio-components/src/components/StudioCheckboxTable/StudioCheckboxTableBody/StudioCheckboxTableBody.tsx
@@ -1,0 +1,12 @@
+import React, { type ReactElement, type ReactNode } from 'react';
+import { StudioTable } from '../../StudioTable';
+
+export type StudioCheckboxTableBodyProps = {
+  children: ReactNode;
+};
+
+export const StudioCheckboxTableBody = ({
+  children,
+}: StudioCheckboxTableBodyProps): ReactElement => {
+  return <StudioTable.Body>{children}</StudioTable.Body>;
+};
