+++ frontend/libs/studio-components/src/components/StudioCheckboxTable/StudioCheckboxTableBody/index.ts
@@ -1,0 +1,1 @@
+export { StudioCheckboxTableBody } from './StudioCheckboxTableBody';
