+++ frontend/libs/studio-components/src/components/StudioCheckboxTable/StudioCheckboxTableHeader/StudioCheckboxTableHeader.module.css
@@ -1,0 +1,3 @@
+.header {
+  background-color: var(--fds-colors-grey-100);
+}
