+++ frontend/libs/studio-components/src/components/StudioCheckboxTable/StudioCheckboxTableRow/StudioCheckboxTableRow.module.css
@@ -1,0 +1,7 @@
+.checkboxCell {
+  padding-right: var(--fds-spacing-8);
+}
+
+.descriptionText {
+  color: var(--fds-semantic-text-neutral-subtle);
+}
