+++ frontend/libs/studio-components/src/components/StudioCheckboxTable/StudioCheckboxTableRow/index.ts
@@ -1,0 +1,1 @@
+export { StudioCheckboxTableRow } from './StudioCheckboxTableRow';
