+++ frontend/libs/studio-components/src/components/StudioCheckboxTable/types/StudioCheckboxTableRowElement.ts
@@ -1,0 +1,6 @@
+export type StudioCheckboxTableRowElement = {
+  value: string;
+  label: string;
+  description?: string;
+  checked: boolean;
+};
